pyproject.toml
[tool.poetry]
name = "investment-tracker"
version = "0.1.0"
description = ""
authors = ["Salvatore Bottiglieri <sasibottiglieri0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openfeed = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
